+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -18,7 +18,7 @@ class KeyConfig {
 
   final bool required;
 
-  final Object? unknownEnumValue;
+  final String? unknownEnumValue;
 
   KeyConfig({
     required this.defaultValue,