+++ json_serializable/lib/src/type_helpers/enum_helper.dart
@@ -3,6 +3,8 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analyzer/dart/element/type.dart';
+import 'package:json_annotation/json_annotation.dart';
+import 'package:source_gen/source_gen.dart';
 import 'package:source_helper/source_helper.dart';
 
 import '../enum_utils.dart';
@@ -44,6 +46,26 @@ class EnumHelper extends TypeHelper<TypeHelperContextWithConfig> {
       return null;
     }
 
+    final jsonKey = jsonKeyForField(context.fieldElement, context.config);
+
+    if (jsonKey.defaultValue == "'${JsonKey.nullForUndefinedEnumValue}'") {
+      throw InvalidGenerationSourceError(
+        '`${JsonKey.nullForUndefinedEnumValue}` cannot be used with '
+        '`JsonKey.defaultValue`.',
+        element: context.fieldElement,
+      );
+    }
+
+    if (!targetType.isNullableType &&
+        jsonKey.unknownEnumValue == JsonKey.nullForUndefinedEnumValue) {
+      // If the target is not nullable,
+      throw InvalidGenerationSourceError(
+        '`${JsonKey.nullForUndefinedEnumValue}` cannot be used with '
+        '`JsonKey.unknownEnumValue` unless the field is nullable.',
+        element: context.fieldElement,
+      );
+    }
+
     String functionName;
     if (targetType.isNullableType || defaultProvided) {
       functionName = r'$enumDecodeNullable';
@@ -53,7 +75,6 @@ class EnumHelper extends TypeHelper<TypeHelperContextWithConfig> {
 
     context.addMember(memberContent);
 
-    final jsonKey = jsonKeyForField(context.fieldElement, context.config);
     final args = [
       constMapName(targetType),
       expression,