+++ json_serializable/test/integration/json_enum_example.g.dart
@@ -8,6 +8,24 @@ part of 'json_enum_example.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
+Issue559Regression _$Issue559RegressionFromJson(Map<String, dynamic> json) {
+  $checkKeys(
+    json,
+    requiredKeys: const ['status'],
+    disallowNullValues: const ['status'],
+  );
+  return Issue559Regression(
+    status: $enumDecodeNullable(_$Issue559RegressionEnumEnumMap, json['status'],
+        unknownValue: JsonKey.nullForUndefinedEnumValue),
+  );
+}
+
+const _$Issue559RegressionEnumEnumMap = {
+  Issue559RegressionEnum.alpha: 'alpha',
+  Issue559RegressionEnum.beta: 'beta',
+  Issue559RegressionEnum.gamma: 'gamma',
+};
+
 const _$StandAloneEnumEnumMap = {
   StandAloneEnum.alpha: 'a',
   StandAloneEnum.beta: 'b',